Refactoring user access functions in access.js
